+++ features/steps/sub_folder1/sub_folder1_steps.py
@@ -1,5 +1,0 @@
-from behave import when
-
-@when('a step is implemented in a sub folder')
-def step_impl(context):
-    context.nested_sub_folder1_visited = True